refactor: rename types->shared_types to avoid collision
Running app.py from g2p/g2p, libraries load g2p.types instead of standard types
and thus app.py was broken.
