+++ g2p/shared_types.py
